[ceres,jxrlib,zziplib] Fixes for non-windows (#6188)
* [ceres, zziplib] fixes for building on non-windows systems

* [jxrlib, zziplib] other fixes for non-windows systems

* [zziplib] fix CMakeLists.txt for non win32 platforms, there were missing exported headers

* [ceres] Fix targets path

* [ceres] Fix typo

* [ceres] trigger CI build

* [ceres] use proper glog/gflags targets

+++ ports/jxrlib/portfile.cmake
@@ -11,9 +11,11 @@ vcpkg_from_github(
 
 file(COPY ${CMAKE_CURRENT_LIST_DIR}/CMakeLists.txt DESTINATION ${SOURCE_PATH})
 
-# The file guiddef.h is part of the Windows SDK,
-# we then remove the local copy shipped with jxrlib
-file(REMOVE ${SOURCE_PATH}/common/include/guiddef.h)
+if(NOT VCPKG_CMAKE_SYSTEM_NAME MATCHES Darwin AND NOT VCPKG_CMAKE_SYSTEM_NAME MATCHES Linux)
+  # The file guiddef.h is part of the Windows SDK,
+  # we then remove the local copy shipped with jxrlib
+  file(REMOVE ${SOURCE_PATH}/common/include/guiddef.h)
+endif()
 
 vcpkg_configure_cmake(
     SOURCE_PATH ${SOURCE_PATH}
